refactor: relocate bank transaction contracts out of component files
Prepares the domain move by making the non-component layers stop depending on
contracts that happened to live in component files. No component moves here, so
tsc alone proves it.

Breaks three import cycles:
- bankTransactionCategoryComboBoxOption.ts -> @internal-types. Pure module, no
  React; providers/BankTransactionsCategorizationStore now imports downward.
- convertApiCategorizationToCategoryOrSplitAsOption -> @utils/bankTransactions/
  categorizationOption. This was a genuine value<->value cycle:
  BankTransactionCategoryComboBox/utils.ts imported hasSuggestions from
  @utils/bankTransactions/shared, which imported this function straight back.
  The other 7 exports in that utils.ts have one consumer and stay nested.
- All four *StringOverrides interfaces -> features/bankTransactions/types.ts.
  BankTransactionsStringOverridesContext imported the type from
  BankTransactions.tsx while BankTransactions.tsx imported the context. Also
  removes two incidental type-only cycles into Header and Table.

Fixes four layer inversions, where @utils/@hooks reached up into @components:
- ExpandedBankTransactionRow/utils.ts -> @utils/bankTransactions/splits.ts
  (9 pure functions over Split[]; unblocks useSplitsForm)
- taxCodeComboBoxOption.ts -> @internal-types (unblocks useTaxCodeOptions)
- DocumentWithStatus -> @internal-types/fileUpload (unblocks hooks/legacy/
  useReceipts). ExpandedBankTransactionRow.tsx declared a byte-identical second
  copy that nothing imported; deleted.

BankTransactions/constants.ts -> features/bankTransactions/constants.ts,
creating the domain root. It was already domain-wide, imported from @utils,
@test-utils, and @views.

No re-export shims: the repo has no barrel files, and BankTransactionsStringOverrides
still reaches dist/index.d.ts structurally through BankTransactionsProps.
