+++ src/components/BankTransactions/RecordManualTransaction/RecordTransactionForm.tsx
@@ -3,6 +3,7 @@ import { type CalendarDate, getLocalTimeZone, today } from '@internationalized/d
 import classNames from 'classnames'
 import { useTranslation } from 'react-i18next'
 
+import { isSplitAsOption } from '@internal-types/bankTransactionCategoryComboBoxOption'
 import type { BankTransaction } from '@internal-types/bankTransactions'
 import { isClassificationExclusion } from '@schemas/categorization'
 import { getDefaultSelectedCategoryForBankTransaction } from '@utils/bankTransactions/shared'
@@ -14,7 +15,6 @@ import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { Form } from '@ui/Form/Form'
 import { Label } from '@ui/Typography/Text'
-import { isSplitAsOption } from '@components/BankTransactionCategoryComboBox/bankTransactionCategoryComboBoxOption'
 import { RecordTransactionFormCategoryCombobox } from '@components/BankTransactions/RecordManualTransaction/RecordTransactionFormCategoryCombobox'
 import { type RecordTransactionFormApi, type RecordTransactionVariant } from '@components/BankTransactions/RecordManualTransaction/useRecordTransactionForm'
 import { FieldErrors } from '@components/forms/FieldErrors'
