+++ src/components/TaxCodeSelect/types.ts
@@ -1,5 +1,5 @@
+import { type TaxCodeComboBoxOption } from '@internal-types/taxCodeComboBoxOption'
 import { type SingleSelectComboBoxProps } from '@ui/ComboBox/types'
-import { type TaxCodeComboBoxOption } from '@components/TaxCodeSelect/taxCodeComboBoxOption'
 
 export type TaxCodeSelectCommonProps = Pick<
   SingleSelectComboBoxProps<TaxCodeComboBoxOption>,
