+++ src/components/features/bankTransactions/types.ts
@@ -1,0 +1,25 @@
+export interface BankTransactionsStringOverrides {
+  bankTransactionCTAs?: BankTransactionCTAStringOverrides
+  transactionsTable?: BankTransactionsTableStringOverrides
+  bankTransactionsHeader?: BankTransactionsHeaderStringOverrides
+}
+
+export interface BankTransactionCTAStringOverrides {
+  approveButtonText?: string
+  updateButtonText?: string
+}
+
+export interface BankTransactionsHeaderStringOverrides {
+  header?: string
+  /** @deprecated Download moved into the header menu and no longer supports a custom label. This override is ignored and will be removed. */
+  downloadButton?: string
+}
+
+export interface BankTransactionsTableStringOverrides {
+  dateColumnHeaderText?: string
+  transactionColumnHeaderText?: string
+  accountColumnHeaderText?: string
+  amountColumnHeaderText?: string
+  categorizeColumnHeaderText?: string
+  categoryColumnHeaderText?: string
+}
