+++ src/providers/BankTransactionsCategorizationStore/utils.ts
@@ -1,5 +1,5 @@
+import type { BankTransactionCategoryComboBoxOption } from '@internal-types/bankTransactionCategoryComboBoxOption'
 import type { SuggestedMatchAsOption } from '@internal-types/categorizationOption'
-import type { BankTransactionCategoryComboBoxOption } from '@components/BankTransactionCategoryComboBox/bankTransactionCategoryComboBoxOption'
 
 export type BankTransactionNonSuggestedMatchOption = Exclude<
   BankTransactionCategoryComboBoxOption,
