+++ src/types/fileUpload.ts
@@ -13,3 +13,13 @@ export type DocumentType =
   | 'PAYROLL_STATEMENT'
   | 'PAYOUT_STATEMENT'
   | 'OTHER'
+
+export interface DocumentWithStatus {
+  id?: string
+  url?: string
+  status: 'pending' | 'uploaded' | 'failed' | 'deleting'
+  type?: string
+  name?: string
+  date?: string
+  error?: string
+}
